Funciona coger la media de los votos

+++ src/main/java/controllers/VerVotos.java
@@ -9,6 +9,6 @@ public interface VerVotos {
 
 	long getNumVotos(int idTema);
 
-    int getValoracionMedia(int idTema, Estudios estudios);
+    double getValoracionMedia(int idTema, Estudios estudios);
 
 }
